refactor: reservation/presentaion 구조 변경
