+++ src/domain/reservation/presentation/api/document/document.decorator.ts
