+++ src/domain/reservation/presentation/api/index.ts
@@ -1,0 +1,2 @@
+export * from './dto';
+export * from './reservation.controller';
