+++ src/domain/reservation/presentation/event/dto/confirm-reservation-event.ts
@@ -1,0 +1,13 @@
+type ConfirmReservationEventProp = { reservationId: number };
+
+export class ConfirmReservationEvent {
+  constructor(readonly prop: ConfirmReservationEventProp) {}
+
+  get reservationId() {
+    return this.prop.reservationId;
+  }
+
+  static from(prop: ConfirmReservationEventProp) {
+    return new ConfirmReservationEvent(prop);
+  }
+}
