+++ src/domain/reservation/presentation/event/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './confirm-reservation-event';
