+++ src/domain/reservation/presentation/event/index.ts
@@ -1,1 +1,2 @@
-export * from './confirm-reservation-event';
+export * from './dto';
+export * from './reservation-event.listener';
