+++ src/domain/reservation/presentation/index.ts
@@ -1,4 +1,2 @@
+export * from './api';
 export * from './event';
-export * from './dto';
-export * from './reservation.controller';
-export * from './reservation-event.listener';
